add some comments for new decoder, move decoder to AddressDecoder and toBitPat api to address set
